+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 -   **dynamic-component-loader** was removed. Use angular's [NgComponentOutlet](https://angular.io/api/common/NgComponentOutlet) instead.
 -   **terra-syntax-editor** has been removed. Use [ck-editor](https://ckeditor.com/ckeditor-4/) instead.
 -   **dependencies** removed peer dependency `ng2-ace-editor` since we no longer depend on it.
+-   **dependencies** changed peer dependency `mydatepicker`to require version 9.
 -   **breadcrumbs-service** removed method initialPath since it has been deprecated.
 -   **terra-tag** removed deprecated `onCloseTag` output.
 -   **terra-taglist** removed deprecated `onCloseTag` output.